Fix methods not using .call() at the end Try to use typescriptUtils.retryNTimes instead of utils.retry3Times as its easier to follow as it expects promises instead of async functions without parameters Move ts types to it's own file. Parameters inside FederatorERC.ts should have a refactor but this should be done later on
