+++ federator/src/contracts/IFederation.ts
@@ -1,6 +1,6 @@
 import { Contract, EventData } from 'web3-eth-contract';
 import { ConfigChain } from '../lib/configChain';
-import { TransactionIdParams, VoteTransactionParams } from './IFederationV3';
+import { TransactionIdParams, VoteTransactionV3Params } from '../types/federator';
 
 export interface IFederation {
   federationContract: Contract;
@@ -16,7 +16,7 @@ export interface IFederation {
 
   hasVoted(txId: string, from: string): Promise<boolean>;
 
-  getVoteTransactionABI(paramsObj: VoteTransactionParams): Promise<any>;
+  getVoteTransactionABI(paramsObj: VoteTransactionV3Params): Promise<any>;
 
   getAddress(): string;
